2020-07-10 MSS Code Factory CFAll 2.13.11195-202007220815 Update version numbers, remanufactured by 2.12.11194-202007220730
The 2.13 projects have been remanufactured by 2.12.11194-202007220730, the version numbers
updated to 2.13.11195, and the Java builds have been completed.  I'm shelving C++ for now; I need
to get something that works done, and a lot of theoretical architectural changes are going to
be implemented for 3.1, so I'm focusing on Java right now.

There were some changes to the XMLNS and URI specifications for the manufactured SAX Loaders,
so rather than go half way, I made the changes to the custom code parsers to match what I changed
in the 2.12 rules.  I think I have everything consistent, which is necessary for deployments.

The msscf_2_13/cartridge-2.13 knowledge base has been refreshed with a snapshot
of the 2.12 knowledge base.  I'll be doing a lot of refactoring of the Java code
to produce 3.1 with the 2.13 knowledge base, but every release starts with
a snapshot of the previous release.  In this case, it will be the end of snapshots
that look like the previous release for very long.  I've a number of code
architecture issues I have gathered over the years that I want to address
for 3.1, including a finer breakdown of the Java packages and the C++ libraries.